[Feat] 캐릭터 움직임 완전 구현
