+++ Assets/Scripts/Crew/Crew.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 93a67780bcd2fb64c87498216b0a4bf1
