+++ Assets/Scripts/Managers/RoomManager.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 7487ae774f38dc548a0f21c667c87196
