+++ Assets/Scripts/Room/Cell.cs
@@ -2,15 +2,5 @@ using UnityEngine;
 
 public class Cell : MonoBehaviour
 {
-    // Start is called once before the first execution of Update after the MonoBehaviour is created
-    void Start()
-    {
-        
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        
-    }
+   public Crew CrewInCell { get; set; }
 }
