+++ Assets/Scripts/Room/Prefab_SelectionSquare.cs
@@ -1,0 +1,5 @@
+using UnityEngine;
+
+public class Prefab_SelectionSquare : MonoBehaviour
+{
+}
