+++ Assets/Scripts/Room/Prefab_SelectionSquare.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 0042db0abe89ab24c88511d330fa5d91
